[#90] refactor: CardViewController 를 파일 분리하고 fileprivate을 private으로 둠 		아직 방법을 못찾아서 internal 접근제한인 것도 있다. 이후 해결할 것이다.
